Reorganize `shapes` module
